Quiver is the core type of this library
Almost everything, but most importantly signatures, will be
parameterized by the type of quivers.  It's sort of like a Carrier
type for upcoming signatures structures.

Furthermore there is a equivalence relation for quiver morphisms and
signature for a isomorphism for that quiver.  I'm not really sure if
_signature of an iso_ makes much sense without the laws, but it
follows the convention of this library to split everything into
signatures and structures.

Also I don't know if that is the best module to hold Equivalence and
isomorphism.