Move Isomorphism into dedicated Morphism modules
I've also added other kind of morphism definitions (Mono and Epi).
Furthermore I've settled for the _⇔_ as the name for the iso
signature.  I think it matches nicely the that it's just the 2 arrows,
there and back, without laws.  I've seen that agda-categories
sometimes rename their _≅_ to the left-right arrow.