Use field name Hom instead of operator for Quiver morphism
This plays nicely with postfix field access notation: `𝒬 .Hom a b`
means arrow `a ⇒ b` in 𝒬.  Sometimes it's necessary to disambiguate
quiver to which the morphism belong, if there are multiple quiver in
the context.  The operator is still a definition in the record module,
one that is defined to be `Hom`.

This let me open publicly `Ob` and `Hom` so I don't have to do it all
the time.  I can use those in postfix form pretty easily, and open
record manually to get nice arrow if I want to.