Move Equivalence from module telescope to record field
I think this might be a bit nicer.  When talking about structures
there always have to be an relevant equivalence instance around.
Moving it away from the module might help a bit with inference, but
that is based on my bad experiences with stuff being in the module
telescope, so...

Anyway it will let me say things like: equivalence used by *this*
structure more naturally.  Perhaps I should change the equivalence from
the instance to regular argument in the homomorphisms?

Speaking of which, when transferring properties via homomorphism, the
resulting structure requires specific equivalence instance.  Now that
is a field I won't have to pass it from the outside as an argument but
just derive it.  I think this'll make nicer interface.