Pass equivalence as regular argument
The rule of the thumb is to have Equivalence be an instance field on
the record/classes, but pass it as regular argument to the functions.

I have some trouble in the past with the equivalence inference.  With
this set up I think it should be fairly explicit which eq to use.